+++ src/Java/myservicebus/src/main/java/com/myservicebus/SnakeCaseEndpointNameFormatter.java
@@ -5,7 +5,7 @@ public class SnakeCaseEndpointNameFormatter implements EndpointNameFormatter {
 
     @Override
     public String format(Class<?> messageType) {
-        String name = messageType.getSimpleName();
+        String name = DefaultEndpointNameFormatter.trimConsumerSuffix(messageType.getSimpleName());
         return name.replaceAll("([a-z0-9])([A-Z])", "$1_$2").toLowerCase();
     }
 }