+++ test/MyServiceBus.Tests/EndpointNameFormatterTests.cs
@@ -8,6 +8,7 @@ namespace MyServiceBus.Tests;
 public class EndpointNameFormatterTests
 {
     class SampleMessage { }
+    class SubmitOrderConsumer { }
 
     [Fact]
     public void Default_returns_type_name()
@@ -21,5 +22,13 @@ public class EndpointNameFormatterTests
     {
         var name = SnakeCaseEndpointNameFormatter.Instance.Format(typeof(SampleMessage));
         Assert.Equal("sample_message", name);
+    }
+
+    [Fact]
+    public void Formatters_trim_MassTransit_endpoint_suffixes()
+    {
+        Assert.Equal("SubmitOrder", DefaultEndpointNameFormatter.Instance.Format(typeof(SubmitOrderConsumer)));
+        Assert.Equal("submit-order", KebabCaseEndpointNameFormatter.Instance.Format(typeof(SubmitOrderConsumer)));
+        Assert.Equal("submit_order", SnakeCaseEndpointNameFormatter.Instance.Format(typeof(SubmitOrderConsumer)));
     }
 }